requirements.txt
absl-py==1.4.0
altair==4.2.0
asgiref==3.6.0
astunparse==1.6.3
attrs==22.2.0
blinker==1.5
cachetools==5.2.1
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
commonmark==0.9.1
contourpy==1.0.7
cycler==0.11.0
decorator==5.1.1
entrypoints==0.4
flatbuffers==23.1.4
fonttools==4.38.0
gast==0.4.0
gitdb==4.0.10
GitPython==3.1.30
google-auth==2.16.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.1
h5py==3.7.0
idna==3.4
importlib-metadata==6.0.0
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
keras==2.11.0
kiwisolver==1.4.4
libclang==15.0.6.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.3
missingno==0.5.1
numpy==1.24.1
oauthlib==3.2.2
opencv-python==4.7.0.68
opt-einsum==3.3.0
packaging==23.0
pandas==1.5.2
patsy==0.5.3
Pillow==9.4.0
plotly==5.12.0
plotly-express==0.4.1
protobuf==3.19.6
pyarrow==10.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydeck==0.8.0
Pygments==2.14.0
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
requests==2.28.2
requests-oauthlib==1.3.1
rich==13.1.0
rsa==4.9
scikit-learn==1.2.0
scipy==1.10.0
seaborn==0.12.2
semver==2.13.0
six==1.16.0
sklearn==0.0.post1
smmap==5.0.0
sqlparse==0.4.3
statsmodels==0.13.5
streamlit==1.17.0
tenacity==8.1.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.11.0
tensorflow-intel==2.11.0
tensorflow-io-gcs-filesystem==0.29.0
termcolor==2.2.0
threadpoolctl==3.1.0
toml==0.10.2
toolz==0.12.0
tornado==6.2
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.14
validators==0.20.0
watchdog==2.2.1
Werkzeug==2.2.2
wrapt==1.14.1
zipp==3.11.0